src/main/java 추가 별점, 댓글 삭제, 주차공간 유무 추가
