+++ ArtConnect/src/main/java/com/multi/artConnect/review/BoardVO.java
@@ -12,6 +12,7 @@ public class BoardVO {
 	private Date reviewDate;
 	private int reviewHit;
 	private int starRating;
+	private boolean parkingAvailable;
 	
 	private String reviewFile;
 	private String filePath;
@@ -112,6 +113,16 @@ public class BoardVO {
 
 	public void setStarRating(int starRating) {
 		this.starRating = starRating;
+	}
+
+
+	public boolean isParkingAvailable() {
+		return parkingAvailable;
+	}
+
+
+	public void setParkingAvailable(boolean parkingAvailable) {
+		this.parkingAvailable = parkingAvailable;
 	}
