+++ ArtConnect/src/main/java/com/multi/artConnect/review/ReplyController.java
@@ -1,5 +1,7 @@
 package com.multi.artConnect.review;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -15,9 +17,9 @@ public class ReplyController {
 
     //댓글 추가 처리
     @RequestMapping(value = "insertReply", method = RequestMethod.POST)
-    public String insertReply(ReplyVO reply, Model model) {
+    public String insertReply(ReplyVO reply, Model model, HttpSession session) {
     	
-    	String memberID = "apple";
+    	String memberID = (String)session.getAttribute("memberID");
 		// 댓글에 사용자의 ID 설정
         reply.setMemberID(memberID);
     	//댓글 추가 
@@ -33,7 +35,7 @@ public class ReplyController {
         return "redirect:/review/boardOne?reviewNO=" + replyVO.getReviewNO();
     }
 
-    @RequestMapping("review/deleteReply")
+    @RequestMapping("deleteReply")
     public String deleteReply(int replyNO, int reviewNO) {
         replyDAO.deleteReply(replyNO);
         return "redirect:/review/boardOne?reviewNO=" + reviewNO;
