--- ArtConnect/src/test/java/com/multi/artConnect/gallery/FilterRequest.java
@@ -1,15 +1,0 @@
-package com.multi.artConnect.gallery;
-
-import java.util.List;
-
-public class FilterRequest {
-    private List<String> filterValues;
-
-    public List<String> getFilterValues() {
-        return filterValues;
-    }
-
-    public void setFilterValues(List<String> filterValues) {
-        this.filterValues = filterValues;
-    }
-}
