--- ArtConnect/src/test/java/com/multi/artConnect/member/MemberController.java.rej
@@ -1,15 +1,0 @@
-diff a/ArtConnect/src/main/java/com/multi/artConnect/member/MemberController.java b/ArtConnect/src/main/java/com/multi/artConnect/member/MemberController.java	(rejected hunks)
-@@ -7,13 +7,9 @@
- 
- import org.springframework.beans.factory.annotation.Autowired;
- import org.springframework.dao.DataIntegrityViolationException;
--import org.springframework.http.HttpStatus;
--import org.springframework.http.ResponseEntity;
- import org.springframework.stereotype.Controller;
- import org.springframework.ui.Model;
- import org.springframework.web.bind.annotation.GetMapping;
--import org.springframework.web.bind.annotation.PostMapping;
--import org.springframework.web.bind.annotation.RequestBody;
- import org.springframework.web.bind.annotation.RequestMapping;
- 
-
