--- ArtConnect/src/test/java/com/multi/artConnect/member/NaverVO.java
@@ -1,65 +1,0 @@
-package com.multi.artConnect.member;
-
-public class NaverVO {
-	private String email;
-	private String nickname;
-	private String name;
-	private String birthday;
-	private String profile_image;
-	private String gender;
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getNickname() {
-		return nickname;
-	}
-
-	public void setNickname(String nickname) {
-		this.nickname = nickname;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getBirthday() {
-		return birthday;
-	}
-
-	public void setBirthday(String birthday) {
-		this.birthday = birthday;
-	}
-
-	public String getProfile_image() {
-		return profile_image;
-	}
-
-	public void setProfile_image(String profile_image) {
-		this.profile_image = profile_image;
-	}
-
-	public String getGender() {
-		return gender;
-	}
-
-	public void setGender(String gender) {
-		this.gender = gender;
-	}
-
-	@Override
-	public String toString() {
-		return "NaverVO [email=" + email + ", nickname=" + nickname + ", name=" + name + ", birthday=" + birthday
-				+ ", profile_image=" + profile_image + ", gender=" + gender + "]";
-	}
-
-}
