--- ArtConnect/src/test/java/com/multi/artConnect/mypage/MypageService.java
@@ -1,5 +1,0 @@
-package com.multi.artConnect.mypage;
-
-public class MypageService {
-
-}
